Revert "feature/screener/exchange-filter"
