--- src/ducks/Watchlists/Widgets/Filter/Metric/Exchanges/index.module.scss
@@ -1,173 +1,0 @@
-@import '~@santiment-network/ui/mixins';
-
-.wrapper {
-  margin-bottom: 24px;
-}
-
-.panel {
-  max-height: 201px;
-  overflow-y: auto;
-  width: 357px;
-}
-
-.item {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  padding: 6px 8px;
-}
-
-.list {
-  padding: 10px 8px;
-}
-
-.list + .list {
-  border-top: 1px solid var(--porcelain);
-}
-
-.action {
-  margin-left: auto;
-  color: var(--waterloo);
-}
-
-.count {
-  color: var(--waterloo);
-  margin-left: 4px;
-}
-
-.name {
-  color: var(--rhino);
-}
-
-.delete {
-  color: var(--waterloo);
-}
-
-.selected {
-  color: var(--casper);
-}
-
-.operator {
-  margin: 4px;
-  color: var(--waterloo);
-
-  @include text('caption', 'm');
-}
-
-.item__selected {
-  cursor: default;
-
-  &:hover {
-    background: none;
-  }
-}
-
-.item:hover {
-  & .action {
-    color: var(--jungle-green);
-  }
-
-  & .selected {
-    color: var(--casper);
-  }
-
-  & .delete {
-    color: var(--persimmon);
-  }
-}
-
-.trigger__btn {
-  min-width: 357px;
-
-  &:hover {
-    border-color: var(--mystic);
-    color: var(--rhino);
-  }
-}
-
-.trigger__arrow {
-  width: 8px;
-  fill: var(--waterloo) !important;
-  transition: transform 0.25s ease-in-out;
-}
-
-.trigger__input {
-  color: var(--waterloo);
-}
-
-.trigger__btn input[data-isactive=true] + .trigger__arrow {
-  fill: var(--rhino) !important;
-  transform: rotate(180deg);
-}
-
-.trigger__btn input::placeholder {
-  color: var(--waterloo) !important;
-}
-
-.dropdown {
-  z-index: 1000;
-}
-
-.loader {
-  height: 32px;
-}
-
-.loader + .loader {
-  margin-top: 2px;
-}
-
-.labels {
-  display: flex;
-  flex-wrap: wrap;
-  margin: 4px 0 6px;
-}
-
-.label {
-  padding: 0 8px;
-  display: flex;
-  align-items: center;
-  margin-bottom: 4px;
-  border: 1px solid var(--porcelain);
-  border-radius: 4px;
-  fill: var(--waterloo);
-  cursor: pointer;
-
-  &:hover {
-    fill: var(--jungle-green);
-  }
-
-  &__close {
-    margin-left: 8px;
-  }
-
-  &__viewMode {
-    cursor: default;
-  }
-
-  @include text('caption');
-
-  &:last-child {
-    margin-right: 0;
-  }
-}
-
-.settings {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.toggleModeBtn {
-  padding-left: 6px;
-  padding-right: 6px;
-  width: 50px;
-  justify-content: center;
-  margin-left: 8px;
-}
-
-.explanation {
-  z-index: 1000;
-  max-width: 190px;
-  padding: 6px 12px;
-}
