+++ src/ducks/Watchlists/Widgets/Filter/Metric/MarketSegments/index.module.scss
@@ -1,9 +1,5 @@
 @import '~@santiment-network/ui/mixins';
 
-.wrapper {
-  margin-bottom: 24px;
-}
-
 .panel {
   max-height: 201px;
   overflow-y: auto;
