+++ src/ducks/Watchlists/Widgets/Filter/Metric/MetricSettings/index.js
@@ -16,6 +16,7 @@ const FilterMetricSettings = ({
   onFirstThresholdChange,
   onSecondThresholdChange,
   onSuggestionClick,
+  autoFocus,
   settings: { firstThreshold, secondThreshold, timeRange, type }
 }) => {
   const inputRef = useRef(null)
