--- src/ducks/Watchlists/Widgets/Filter/Metric/utils.js
@@ -1,18 +1,0 @@
-export function filterValuesBySearch (value = '', values, key) {
-  if (!value) {
-    return values
-  }
-
-  const chars = value.toLowerCase().split('')
-  const passedValues = []
-
-  values.forEach(item => {
-    const str = item[key].toLowerCase()
-    const foundChars = chars.filter(char => str.includes(char))
-    if (foundChars.length === chars.length) {
-      passedValues.push(item)
-    }
-  })
-
-  return passedValues
-}
