+++ src/ducks/Watchlists/Widgets/Filter/dataHub/metrics.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { Filter } from './types'
 import MarketSegments from '../Metric/MarketSegments'
-import Exchanges from '../Metric/Exchanges'
 import {
   percentValueFormatter,
   percentServerValueFormatter,
@@ -14,8 +13,7 @@ import { formatNumber } from '../../../../../utils/formatting'
 import {
   ETH_SPENT_CELL,
   MARKET_SEGMENTS_CELL,
-  RANK_CELL,
-  EXCHANGES_CELL
+  RANK_CELL
 } from '../../Table/Columns/columns'
 
 export const METRIC_PERCENT_SUFFIX = '_change_'
@@ -583,14 +581,6 @@ export const Metric = {
     shortLabel: 'Whale T.C>1m$',
     aggregation: AGGREGATIONS_LOWER.SUM,
     showTimeRange: true
-  },
-  traded_on_exchanges: {
-    category: CATEGORIES.FINANCIAL,
-    label: 'Traded on Exchanges',
-    accessor: 'tradedOnExchangesCount',
-    isStatic: true,
-    Cell: EXCHANGES_CELL,
-    Widget: props => <Exchanges {...props} />
   }
 }
 
@@ -668,8 +658,7 @@ export const metrics = [
   Metric.age_consumed,
   Metric.network_profit_loss,
   Metric.whale_transaction_count_100k_usd_to_inf,
-  Metric.whale_transaction_count_1m_usd_to_inf,
-  Metric.traded_on_exchanges
+  Metric.whale_transaction_count_1m_usd_to_inf
 ]
 
 export const MetricAlias = {
