+++ src/ducks/Watchlists/Widgets/Table/Columns/columns.module.scss
@@ -5,15 +5,6 @@
   color: var(--rhino);
 }
 
-.exchanges {
-  display: inline-flex;
-  width: fit-content;
-
-  &__count {
-    margin: 0 6px;
-  }
-}
-
 .segments {
   display: inline-flex;
   width: fit-content;
