+++ src/ducks/Watchlists/Widgets/TopPanel/index.module.scss
@@ -21,7 +21,7 @@
 }
 
 .open {
-  z-index: 26;
+  z-index: 24;
   position: absolute;
 }
