+++ src/pages/Index/Section/Personal/index.module.scss
@@ -70,7 +70,7 @@
   &_active {
     background-color: var(--jungle-green);
     border-color: var(--jungle-green);
-    fill: $white;
+    fill: var(--white);
   }
 }
